requirements.txt
langchain
langchain-openai
langchain-huggingface
langchain-chroma
chromadb
gradio
pypdf
sentence-transformers
python-dotenv